# API-OPENAPI-SPEC.yaml
openapi: 3.0.0
info:
  title: Herald Social API
  description: Complete REST API specification for Herald Social mobile application
  version: 1.0.0
  contact:
    name: Herald Team
    url: https://heraldsocial.com
  license:
    name: MIT
servers:
  - url: https://herald-backend-ra5x.onrender.com/api/v1/
    description: Production server
  - url: https://staging-api.heraldsocial.com/v1
    description: Staging server
  - url: http://localhost:8000/v1
    description: Development server
components:
  securitySchemes:
    Bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token for authentication
  schemas:
    UUID:
      type: string
      format: uuid
      example: "550e8400-e29b-41d4-a716-446655440000"
    User:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        user_id:
          $ref: '#/components/schemas/UUID'
        username:
          type: string
          example: "johndoe"
        display_name:
          type: string
          example: "John Doe"
        full_name:
          type: string
          example: "John Michael Doe"
        email:
          type: string
          format: email
          example: "john@example.com"
        avatar_url:
          type: string
          format: uri
          nullable: true
        bio:
          type: string
          nullable: true
          example: "Software engineer and tech enthusiast"
        tier:
          type: string
          enum: [free, creator, premium]
          default: free
        reputation:
          type: integer
          example: 8500
        is_verified:
          type: boolean
          default: false
        is_creator:
          type: boolean
          default: false
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
paths:
...
  /users/me:
    get:
      tags:
        - Users
      summary: Get current user profile
      security:
        - Bearer: []
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized

    patch:
      tags:
        - Users
      summary: Update current user profile
      security:
        - Bearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                display_name:
                  type: string
                bio:
                  type: string
                avatar_url:
                  type: string
                  format: uri
      responses:
        '200':
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized

    delete:
      tags:
        - Users
      summary: Delete current user account
      security:
        - Bearer: []
      responses:
        '204':
          description: Account deleted
        '401':
          description: Unauthorized
...
security:
  - Bearer: []
...